[Gradle|Sync] cleanup: moved dependency resolver test to dependencyModel package
This is a follow-up to commit c4be347449156b43b1fb10b7b1b671ea4538525a that moved the resolver class itself, and allows reducing the visibility of method `chooseAuxiliaryArtifactFile` to package private again
