+++ plugins/gradle/tooling-extension-impl/testSources/com/intellij/gradle/toolingExtension/impl/model/dependencyModel/GradleDependencyResolverTest.kt
@@ -1,7 +1,6 @@
 // Copyright 2000-2024 JetBrains s.r.o. and contributors. Use of this source code is governed by the Apache 2.0 license.
-package org.jetbrains.plugins.gradle.tooling.util.resolve
+package com.intellij.gradle.toolingExtension.impl.model.dependencyModel
 
-import com.intellij.gradle.toolingExtension.impl.model.dependencyModel.GradleDependencyResolver
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.io.TempDir
